Updates
Got rid of some database migrations that were messing with the authentication feature - still need to work on cleaning that up as well but we can go ahead and start working out the schedule-generating logic now. Also updated the requirements.txt file since I had to download a mySQL database connector to connect the project to the database.